python/pyproject.toml
[project]
name = "actiontab-server"
version = "0.1.0"
description = "MLX-VLM prediction server for ActionTab"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pillow>=10.0.0",
    "mlx>=0.21.0",
    "mlx-vlm>=0.1.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]

[dependency-groups]
dev = []

